Update tests, rename crud fixture
