+++ tests/e2e/fixtures/d1_crud/cidl.pre.json
